pyproject.toml
[project]
name = "llm-caller"
version = "0.1.0"
description = "A Python application for calling various LLM APIs (OpenAI, Gemini, Azure OpenAI)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.50.0",
    "google-generativeai>=0.8.5",
    "openai>=2.7.2",
    "python-dotenv>=1.2.1",
]
